services/manager/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.1
pydantic==2.10.4
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-multipart==0.0.19
aiokafka==0.12.0
redis[hiredis]==5.2.1
elasticsearch[async]==8.17.0
httpx==0.28.1
websockets==14.1
python-dateutil==2.9.0
orjson==3.10.13
structlog==24.4.0
